feat: split domain trials dashboard into separate Start and Status pages
The 872-line DomainTrialsDashboard combined launch controls, status
monitoring, and backfill into one page. This splits it into:

- DomainStartBatches (345 lines) — launch form with backfill mode via
  URL params (?evaluationId=&models=&depth=)
- DomainStatus (262 lines) — monitoring with polling and fill-gap links

Deletes BackfillConfirmModal and the old backfill UI. All files under
the 400-line CI limit.
