+++ cloud/apps/web/src/pages/Dashboard.tsx
@@ -86,7 +86,7 @@ export function Dashboard() {
           </p>
           <div className="mt-4 grid gap-3 md:grid-cols-3">
             <Link
-              to={`/domains/${activeDomain.id}/run-trials`}
+              to={`/domains/${activeDomain.id}/status`}
               className="rounded-lg border border-gray-200 bg-gray-50 p-4 hover:bg-gray-100"
             >
               <div className="text-xs font-medium uppercase tracking-wide text-gray-500">Resume batches</div>
