+++ cloud/apps/web/src/pages/Domains.tsx
@@ -70,7 +70,7 @@ export function Domains() {
             </select>
             {selectedDomain != null && (
               <Link
-                to={`/domains/${selectedDomain.id}/run-trials`}
+                to={`/domains/${selectedDomain.id}/start`}
                 className="inline-flex items-center justify-center rounded-lg border border-teal-600 bg-teal-600 px-4 py-2 text-sm font-medium text-white hover:bg-teal-700 focus:outline-none focus:ring-2 focus:ring-teal-500 focus:ring-offset-2 md:ml-auto"
               >
                 Add Paired Batches for all Vignettes
