+++ cloud/apps/web/tests/pages/Dashboard.test.tsx
@@ -22,7 +22,7 @@ describe('Dashboard', () => {
     expect(screen.getByRole('link', { name: /validation/i })).toHaveAttribute('href', '/validation');
     expect(screen.getByRole('link', { name: /archive/i })).toHaveAttribute('href', '/archive');
     expect(screen.getByRole('heading', { name: /resume active domain work/i })).toBeInTheDocument();
-    expect(screen.getByRole('link', { name: /open domain level batches/i })).toHaveAttribute('href', '/domains/domain-a/run-trials');
+    expect(screen.getByRole('link', { name: /open domain level batches/i })).toHaveAttribute('href', '/domains/domain-a/status');
     expect(screen.getByRole('link', { name: /open findings for this domain/i })).toHaveAttribute('href', '/domains/analysis?domainId=domain-a');
     expect(screen.getByRole('link', { name: /review setup coverage/i })).toHaveAttribute('href', '/domains?domainId=domain-a&tab=setup&setupTab=contexts');
   });
